Initial Release of Live Light
First release of live light very simple logic. The server allow us to
query the current colour and update it. The client will update the
colour from green to red and vice versa. The client makes a request when
a USB is plugged in or out.

IRL it changes when I plug in my headset, red, and green when it's
plugged out.

Fixes #1